Emprestimo controller

+++ miniEstante/src/miniEstante/model/Emprestimo.java
@@ -2,37 +2,44 @@ package miniEstante.model;
 
 public class Emprestimo {
 
-	long idCliente;
-	long idLivro;
+	int idCliente;
+	int idLivro;
 	String dataAluguel;
 	String dataDevolucao;
 	
 	
-	public Emprestimo(long idCliente, long idLivro, String dataAluguel, String dataDevolucao) {
+	public Emprestimo(int idCliente, int idLivro, String dataAluguel, String dataDevolucao) {
 		this.idCliente = idCliente;
 		this.idLivro = idLivro;
 		this.dataAluguel = dataAluguel;
 		this.dataDevolucao = dataDevolucao;
 	}
 	
-	
-	public long getIdCliente() {
+
+
+	public int getIdCliente() {
 		return idCliente;
 	}
 
-	public void setIdCliente(long idCliente) {
+
+
+	public void setIdCliente(int idCliente) {
 		this.idCliente = idCliente;
 	}
 
-	public long getIdLivro() {
+
+
+	public int getIdLivro() {
 		return idLivro;
 	}
 
 
 
-	public void setIdLivro(long idLivro) {
+	public void setIdLivro(int idLivro) {
 		this.idLivro = idLivro;
 	}
+
+
 
 	public String getDataAluguel() {
 		return dataAluguel;
